requirements.txt
requests>=2.28.0
pydantic>=1.10.0
python-dotenv>=1.0.0
praw>=7.7.0
python-dateutil>=2.8.2
